feat: actualización definitiva del módulo de gastos

+++ Spendy/src/main/java/com/example/Spendy/repositorios/IGastoRepositorio.java
@@ -1,16 +1,10 @@
 package com.example.Spendy.repositorios;
 
+import com.example.Spendy.modelos.Gasto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.Spendy.modelos.Gasto;
-
-@Repository 
-public interface IGastoRepositorio extends JpaRepository <Gasto, Integer>{
-
-       //Guardar
-    //Buscar por Id
-    //Buscar todos los registros
-    //Modificar por Id 
-    //Eliminar por Id 
+@Repository
+public interface IGastoRepositorio extends JpaRepository<Gasto, Long> {
+    // Proporciona los métodos para persistir los consumos o gastos financieros
 }
